Hits needed to kill an enemy is a float, instead of int in order to enable attack bonuses if full power bar is used + removed regenerating enemy

+++ Prototype_1/Assets/Scripts/Misc/GameManager.cs
@@ -77,7 +77,7 @@ public class GameManager : MonoBehaviour
         // TODO: save as persistent the new value of maximum linkable stars
     }
 
-    public float GetEnemyHealthFromHits(int hitsToDeath)
+    public float GetEnemyHealthFromHits(float hitsToDeath)
     {
         return hitsToDeath * enemyPerStarDamage;
     }
